[Telemetry] Simplify TraceValue attributes
This is the first in a series of CLs to trim down and simplify the API
of TraceValue objects.

To start we reduce the number of arguments required by the constructor,
allowing the TraceValue itself to compute derived attributes.

When all clients switch to using the new constructor arguments, their
old computation can be removed from TelemetryInfo (where they didn't
really belong).

Bug: chromium:928278
Change-Id: I05a71b8efaefe00858dcd445220ca3549c57d0ea