Reland "speed: Look only for the appropriate devices/browsers."
This change has gone through several land/reverts:
 . landed in crrev.com/c/1474597, reverted in crrev.com/c/1481933
 . landed in crrev.com/c/1481961, reverted in crrev.com/c/1487912
 . landed in crrev.com/c/1487555, reverted in crrev.com/c/1493430

The latest reverts happened because the bots were attempting to run
benchmarks for unsupported platforms. This has since been fixed in
crrev.com/c/1493202, so this should no longer be an issue. Parts of the
original CL also landed in crrev.com/c/1494782. The rest can now land.

Original CL description:

When running a benchmark, look only for the devices/browsers that
are supported by the benchmark. For example, when running a desktop
benchmark (e.g. rendering.desktop), the code still tries to discover
the android devices attached to the workstation (which takes ~2
seconds), and then the list of browsers appropriate for that device
(which can take ~10 seconds, if you have an android device attached).
This is unnecessary, since these android devices or browsers are not
going to be used anyway. So avoid looking for these devices/browsers.
This brings down the runtime at startup by ~15 seconds. So running a
single story goes down from ~45 seconds to ~30 seconds.

BUG=chromium:932409
BUG=chromium:935578

Change-Id: Ia1dc7b884cd2e6f93ead598c3ca57b2ecfad51cd