Fix spa tests.
The new version of chromium bundled with WCT no longer supports html
imports and shadow dom v0, so these tests are falling back to the
polyfill.
This CL updates a createShadowRoot() call, fixes a bug in the html
imports polyfill, and fixes some timing issues around cp.measureText()
in order to make the tests pass again.
When v2spa is merged to master branch, I'll update it to lit-element
and translate all html imports to es6 modules so we won't need the
polyfill.

Bug: chromium:918193
Change-Id: I8537145b14f65050e0e99127d2cca7017a2f3480