Update patch set 4
Patch Set 4:

(3 comments)

Hmm.. I left a few comments below. But after looking at the hackyness of this, and noticing this wont play very well with my idea of allowing the user to override the --browser from the extra args in the pinpoint forms (because it will be overriden again at this lower run level), I'm not sure any more if this is the best approach.

So sticking with the same browser for all runs as you had also suggested is probably best. Pinpoint won't be able to flag the browser flipping change as culprit of any regressions (so perf-sheriffs need to be very well aware of that, maybe also show a warning in pinpoint?), but at least we will be able to find regressions on the new --browser android-chrome which happen near the flip.

Wdyt?

Patch-set: 4
Reviewer: Gerrit User 1123010 <1123010@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>