Introduce SetPossibleBrowser method to SharedState
Currently FindBrowser() is called several times from different parts
of code. To simplify and speed up the code, we want to construct a
PossibleBrowser object only once and pass it everywhere.
First step is to pass it to SharedPageState. For compatibility, we
leave the constructor interface intact, but add a function to set
the possible browser.

Bug: chromium:939374
Change-Id: Iff6d50ac887f1bc9f9e7bb8686cc208365249baf