Update patch set 1
Patch Set 1:

Discussed this with Juan offline, decided to introduce a new parameter to constructor of SharedState. The process of adding this parameter will be more complicated, but we will end with a more intuitive interface in the end.

I abandon this CL in favor of the other approach.

Patch-set: 1
Reviewer: Gerrit User 1328067 <1328067@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>