Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> > Patch Set 7:
> > 
> > jank_count is an old metric that has existed for a while on .mobile benchmark. Are we sure we want to change the existing metric?
> > 
> > +vmiura@ for thoughts.
> 
> I think it's a good change. Having a higher jank_count doesn't necessary mean the page is janky with our current set up (ex: the recorded animation may be too long).

I agree. jank/second also makes it easier to compare the metric across pages. But the duration of the animation typically shouldn't change between runs unless we refresh the page (which happens rarely), and changing the definition of the metric often makes it difficult for bisect/dashboard etc. So if we were to re-define the existing metric, we would need to be careful with how we go about doing this.

Patch-set: 7
Reviewer: Gerrit User 1000111 <1000111@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>