Use incremental file decoding and JSON parsing
Before this change, we decompressed the whole compressed histograms file
in memory, then parsed the whole JSON object that way. This change
implements a couple of things which makes require less RAM:

- A 'DecompressFileWrapper' which handles incremental file data reading
and on-the-fly zlib decompression. This does inline buffering, and
allows us to cap the amount of memory we use when reading decompressed
data.

- A helper function '_LoadHistogramList' which uses the JSONDecoder type
for incrementally decoding JSON from fragments we're loading from a
file being decompressed on the fly.

In this change we also update the mock used to simulate a file hosted in
Google Cloud Storage with a Fake, which behaves like a file (adding
write and read methods) but still supports being inspected as-if a mock
object.

R=benjhayden@chromium.org, simonhatch@chromium.org

Bug: chromium:917914
Change-Id: Id1141ce05e19effdf20d73a16bbd64ef4c716084