Allow pages to configure which metrics should be run.
After the benchmarks merge (http://bit.ly/why-merge-benchmarks), we ended
up with some benchmarks that have certain pages that apply to

If we landed this changelist, then rendering team could do the following
to fix this.

1. Break up rendering_metric.html into
 - rendering_frametime_metric.html
 - rendering_latency_metric.html
 - rendering_pipeline_metric.html
 - rendering_pixels_metric.html
 - rendering_queueing_duration_metric.html
(For example, I'm not sure what the actual break down
would be).

2. Then edit
https://cs.chromium.org/chromium/src/tools/perf/benchmarks/rendering.py?l=43
to remove renderingMetric for it, and edit the pages in
https://cs.chromium.org/chromium/src/tools/perf/page_sets/rendering/
to override AddTracingMetrics().

Bug: chromium:906509, chromium:909961, chromium:942196
Change-Id: I59e22a5d8012a6683407cdc13f6f69023dd3a3b8