Use ijson for proper incremental parsing
The previous implementation of the incremental parsing actually doesn't
support partial parsing of JSON elements. This change makes it so that
we use ijson to process files incrementally and build objects as they
come.

Future work should allow us to turn this into a stream processing
implementation, which can handle histograms as they are parsed instead
of building a list in memory of the dicts before parsing.

R=simonhatch@chromium.org

Bug: chromium:941888, chromium:942368
Change-Id: I9f94341c4d2d7c766fd1d10d684e0167227059d3