Fix ecmaVersion, symlink in deploy.
Spa currently uses Polymer 2.0, vulcanize, and some ES9 (2018) features like
async generators.
However, vulcanize was deprecated shortly after ES6, so it doesn't automatically
support ES9, and there doesn't seem to be an easy way to configure ecmaVersion.
This CL adds a sed command to the dashboard's deploy script to bump vulcanize's
ecmaVersion to 9.
This CL also adds a few symlinks to help appengine find all the necessary
third_party libraries.

With these changes, a clean checkout (with a small change to app.yaml) can deploy spa.

After spa launches, I plan to migrate it to lit-element and replace vulcanize
with webpack, so both these hacks will not be necessary. That migration can
be partly scripted so it should only take a couple days.

Bug: chromium:918193
Change-Id: I268137a792fbdea93b213984db85bbe9833f11ab