[Telemetry] Only print detailed browser into for first test run.
The commandline arguments that Chrome uses are extremely long,
and they cause us to hit a 16Mb log limit.
With this limit hit, it because difficult to figure out
the root cause of hangs and failures since those logs are missing.
These logs also generally slow down loading of the UI and make it
difficult to find errors.

This change is a no-op for the typ-based Telemetry tests (although
it does add an argument that they could take advantage of if they want)
. For the Telemetry benchmarking tests, this causes us to only print
the commandline arguments to start Chrome and the browser info log
for the first story inside a benchmark.

Bug: chromium:943650
Change-Id: I34c359332fc2b0631b6490f0463a8b26df0dafec