[Telemetry] Only print detailed browser into for first test run.
The commandline arguments that Chrome uses are extremely long,
and they cause us to hit a 16Mb log limit.
With this limit hit, it becomes difficult to figure out
the root cause of hangs and failures since those logs are missing.
These logs also generally slow down loading of the UI and make it
difficult to find errors.

This change is a no-op for the typ-based Telemetry tests (although
it does add an argument that they could take advantage of if they want)
. For the Telemetry benchmarking tests, this causes us to print
the commandline arguments to start Chrome and the browser info log
for the first story inside a benchmark and thereafter to print shortened
versions. In particular, for commandline, we drop the finch-related
flags and replace them with "...".

This change replaces
https://chromium-review.googlesource.com/c/catapult/+/1530028
(which didn't work: I didn't test it, which was a mistake.)

It also fixes up
https://chromium-review.googlesource.com/c/catapult/+/1531402
which switched commandline second print to be only done in
debug mode. With this change, we print it at info level but
only for the first test run.

Bug: chromium:943650
Change-Id: I34c359332fc2b0631b6490f0463a8b26df0dafec