Implement option for TabListBackend.New to open in a new window
The new window is opened via the JavaScript window.open() function, but
ideally the DevTools protocol would be extended to respond to requests
for tabs to be opened in new windows: crbug.com/943279.

An alternate way to force the navigation toolbar, etc. to be included
would be to add Ctrl and N mappings to key_event.py and sending both.

This enables Telemetry tests with multiple windows.

Change-Id: I4b607bfb4b7dbf482c693251ea784560242c5f42
Bug: chromium:706860