Fixes issue where assertion errors are raised when tests are added to a test set
When unit tests co-exists in a directory with telemetry browser tests, an assertion
error is raised when attempting to run a telemetry browser test. This happens because
Typ attempts to classify the unit tests in the directory also. In http://crrev.com/c/1529640,
we add an assert statement that checks if the test's fully qualified name starts with
the test name prefix. This CL will replace the assert statement with an if statement that
will classify the test only if it's fully qualified name starts with the test name prefix.
If it does not start with the prefix and if the verbosity level is greater than 3 then
a message stating that the test does not start with the prefix and will not run will be
printed.


Change-Id: Iaee05865e8914c28f86a8a0c852e9516324f8101
Bug: chromium:698902