Update patch set 3
Patch Set 3:

Hi Dirk

I noticed an issue when telemetry unit test scripts exist in the same top level directory of telemetry browser tests scripts. Typ tries to classify the tests in the unit test script as well as the browser test script. From a previous CL, we raise an assertion error when the test id does not start with the prefix. In this CL I try to replace that with something else. If the test fully qualified name does not start with the test name prefix, then it will not classify it and if the verbosity is set to 3 then it will output an error message stating that test will not be run because the fully qualified name does not start with the test name prefix passed at the command line. Can you take a look.

Best Regards
Rakib hasan

Patch-set: 3
Reviewer: Gerrit User 1003232 <1003232@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1311766 <1311766@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false