Update patch set 3
Patch Set 3:

> Patch Set 2: Code-Review+1
> 
> oh, sorry, I misread things. Rakib is removing this, not Caleb. 
> 
> That's somewhat better, but do we know why this is flaky?

The test depends on the discovery mechanism to find tests with different prefixes in the top level directory that was set. Some things can get removed and cause the test to only discover tests with the same prefix and not raise an assertion. This was a bad test from the beginning to push and its good we removed it.

Patch-set: 3
Reviewer: Gerrit User 1311766 <1311766@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>