Write telemetry traces in proto format: standalone version
There are some difficulties deploying all the necessary libraries to
all the bots, so let's have a look at another option: writing protobufs
directly without reliance on external libraries and compiled classes.

I believe it will be harder to debug and support, but perhaps we can
start with this version and switch to using proper libraries when
all the issues are solved.

Bug: chromium:944418
Change-Id: I14987010b19f409b01e8f3bf627c0cba06506d06