Write telemetry traces in proto format: standalone version
There are some difficulties deploying all the necessary libraries to
all the bots, so let's have a look at another option: writing protobufs
directly without reliance on external libraries and compiled classes.

I believe it will be harder to debug and support, but perhaps we can
start with this version and switch to using proper libraries when
all the issues are solved.

Another small difference from the previous CL [1] is delta-encoding of
timestamps. Everything else must be the same.

[1] https://chromium-review.googlesource.com/c/catapult/+/1524582

Bug: chromium:944418
Change-Id: I032add105d33c0c69f2cb5847185a06341e45630