[Telemetry] Cleanup TraceDataBuilder in CPU tracing agent tests
With the new TraceDataBuilder API, clients now need to ensure to clean
up their written trace data.

We keep the usage of AsData() to read the written trace data for this
tests, for now this is the right place to keep these tests since the
code for this tracing agent formatting the trace data lives in the
corresponding cpu_tracing_agent.py.

Future work might move the trace formatting code for this agent, and
corresponding tests, out of Telemetry.

Bug: chromium:928278
Change-Id: I2e21f91b4fd95448eb1dc9c2c2c9b65cf9c8f501