[tracing] Restrict writing unstructured trace data
As part of Chrometto work to modernize the tracing system and switch to
proto based format, we want to limit the kinds of tracing data that can
be written with TraceDataBuilder.

As a start, we will only allow writing unstructured text data (as done
e.g. by atrace tracing agents) if explicitly requested.

Bug: chromium:928278
Change-Id: Ie0b44a9e6bc2436b33bd63ef472232abf3b08092