[Typ] Adds new function to Typ's Runner class
This function will take a test case object and then get the expectations
for the test. It will remove the test name prefix from the test case's id
and then find search the test expectations for that string. This function
will also update SeriallyExecutedBrowserTestCases's GetExpectationsForTest
to call this new function.

Bug: chromium:698902
Change-Id: I7d1ae6f04223c8f908054876fa690b0b45307b76