[Telemetry] Reuse devtools http and websocket connections
To check whether a devtools agent is ready, Telemetry used to attempt
establishing http and websoclet conections and, if successful,
disconnect from them. Later the devtools clients would re-connect to
those.

This, however, would no longer work after new changes in Chrome
tracing; since the first browser connection would adopt the tracing
session, which is then lost when disconnecting.

With this change, if successful, we keep the established connections
used when waiting for the agent to become ready, and re-use them for
the actual devtools client created.

Bug: chromium:945876
Change-Id: I74fd2f8baf1cc2511a4e4ed09157a952138c3aa4