[Telemetry] Reuse devtools http and websocket DevTools connections
To check whether a DevTools agent is ready, Telemetry used to attempt
establishing http and websocket connections and, if successful,
disconnect from them. Later the DevTools client would re-connect
again to those.

This, however, will no longer work after new changes in Chrome
tracing; since the first browser connection would adopt the tracing
session, which is then lost when disconnecting.

With this change, if successful, we keep the established connections
used while waiting for the agent to become ready, and re-use them for
the actual DevTools client created.

Bug: chromium:945876
Change-Id: I74fd2f8baf1cc2511a4e4ed09157a952138c3aa4