crappy client-side TeamFilter
This is a first attempt at an implementation for TeamFilter, which would allow
teams to use e.g. #team=fuchsia to filter the menus of Suites, Sheriffs, Reports, etc.
This implementation is purely client-side, so it uses hashes to obscure
internal-only names. The hashes are difficult to manage.
A better approach would involve modifying the API to accept e.g. ?team=fuchsia
and use namespaced_stored_objects so that teams can manage explicit lists of
Suites etc in the datastore. The service worker CacheRequest classes would also
need to add the team name to their database keys.

In the UI, a team picker widget might fit either in the header or the drawer.

Change-Id: I91de90aa298af685cdff8fcd907960b43b4f36b5