Reland: Add dont-require-rooted-device compatibility-mode to Telemetry.
Relanding with SELinux sandbox disabled except under this
compatibility mode.

This avoids doing all of the operations which require root access. It
switches the command line file to the version in /data/local/tmp, and
also places the user profile in a directory under /data/local/tmp.

Tested locally by running a subset of the WebGL conformance tests on a
non-rooted Pixel 2 with Chrome Canary:

  ./content/test/gpu/run_gpu_integration_test.py webgl_conformance \
    --browser=android-chrome-canary \
    --compatibility-mode=dont-require-rooted-device \
    --test-filter=gpu_tests.\
    webgl_conformance_integration_test.WebGLConformanceIntegrationTest.\
    WebglConformance_conformance_attribs\*

Bug: chromium:826295
Change-Id: Icc79b9435e387aecfbb9709a424225e1f14c7c0c