Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> lgtm thanks!
> 
> how did you choose the values? Are they basically the closest to N5X or there was some deeper investigation? Any info how much it is expected to reduce the noise?
> 
> ps: this will likely move our metrics across the board, so better be done at the time of low commit volume (a weekend?), with a note to perf-sheriffs@chromium.org

Thanks for the reply :)
So the values were chosen because the N5/N5X values seem to be approximately half of the big cores' max frequency. Prior to this change, speedometer2 (RunsPerMinute) reported a standard deviation of 5.1 and a CV of 0.12, with this change I am observing a standard deviation of 1.07, and a CV of 0.04. 

Is this enough? Would you need anymore investigation?

Patch-set: 1
Reviewer: Gerrit User 1338105 <1338105@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>