Update patch set 1
Patch Set 1:

it's ok. While we could be more thorough, our time is oimited, and this is imo a reasonable state to be in. I'm calling for details in order for it to be easier for the next person to pick it up.

Patch-set: 1
Tag: mailMessageId=<CAH3q7_mbyx_tLmKgHpD8PevTNsPGgtTvyn3ZsLX3vYkYrBUJ2Q@mail.gmail.com>