Pinpoint - Look for performance_browser_tests in browser_tests
The output from performance_browser_tests is currently being output
in browser_tests, and so every pinpoint job fails thinking there was
no output. Until this is fixed properly, if retrieving the output
fails, look in the other name. Eventually this should just work
properly and we can remove the hack.

Bug: chromium:947501
Change-Id: I7c2b55418fe11a719116ea5806523c31659edc00