[Telemetry] Support bundles in Telemetry tests
Adds support for using bundle versions of Chrome (specifically
Chrome Modern) in Telemetry tests. The ability to run Telemetry using
bundles was added previously, but required using the --browser=exact
argument.

Additionally, adds support for installing modules besides the default
"base" module, which is required for testing features such as VR that
are not part of the base module.

Bug: chromium:942132
Change-Id: I09d7ad0148292722cbadc0812116ae4da05418c5