Update patch set 4
Patch Set 4:

The basic idea is good here, but the way this is implemented, all of the fields will be written as top-level keys, and I think that gets increasingly clunky.

Can we create a top-level 'metadata' key and write things under that?

We'll probably need a whitelist of values that we should keep writing to the top-level, at least for now, as well.

Patch-set: 4
Reviewer: Gerrit User 1003232 <1003232@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>