Remove restriction that a test expectation glob can only end with a '*'
In gpu tests some tests are made from concatenating a base name with a suffix
which represents an html page. There may be several tests with the same
suffix but different base names. By removing the restriction of having the
'*' at the end of a glob, people can write one expectation for a suffix
where the '*' takes place of the base name as a opposed to writing a test
expectation for every base name

Bug: chromium:698902
Change-Id: I57b5a0d6d04847ef2f0fb91aa029ccdaf634288d