Telemetry: separate the preparation code while keeping old interface
This is a reland of https://chromium-review.googlesource.com/c/catapult/+/1508282.
The differnce is that we keep all methods that are used in chrome tests.
I marked them as deprecated in the code.
When (if) we move telemetry to chromium repository, we will be able to remove them.

Bug: chromium:939374
Change-Id: Ib93a102e5d199d91a3a8ea269c6ca9c1bdfa4349
TBR: crouleau@chromium.org