Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> Sorry, nevermind, I forgot that drawTrack isn't tested.
> I tested this manually locally and it seems to work!
> Would it be possible to fix this where flow arrows are drawn?
> I see a TODO in drawFlowArrow_ pointing to issue 931. Is that related?

Thanks for the feedback! I've added a few lines to account for the translation that might have been applied right before the arrow is actually drawn. We could also restore and re-save the context at the beginning of drawFlowArrows_ or save the value in ctx.translate() and pass it into that function if that's cleaner. Just let me know what you'd prefer.

Issue 931 seems to be referring to drawing flow events when a track is collapsed: https://github.com/catapult-project/catapult/issues/931. I think that would be a separate fix to get that working.

And lastly thanks for working on about:tracing in general, it's been super helpful!

Patch-set: 2
Reviewer: Gerrit User 1323689 <1323689@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>