Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Simon, can you please re-deploy the ChromePerf/Pinpoint app for this change and let me know once this is done? This is blocking landing changes in https://crrev.com/c/1536034. Thanks.

Friendly ping. I'd also be happy to re-deploy the app myself if you point me to the instructions. I know the usual process of uploading/switching versions on AppEngine, but not sure whether there are some additional non-standard steps that I should follow and which metrics to watch after deployment.

Patch-set: 3