Revert "Adds optional expectations_files and tags fields to the JSON results"
This reverts commit fa7f5cb5c32382af9e39f8f676d3b4ef6c81d5fa.

Reason for revert: There are failures in the catapult auto roll because I have
some tests in the GPU tests folder which assume that test-name-prefix is in the
top level of the JSON. However in this CL I moved them into the metadata.
So those tests are failing.

Original change's description:
> Adds optional expectations_files and tags fields to the JSON results
>
> These fields can be used to debug regressions due to invalid test
> expectations.
>
> Bug: chromium:698902
> Change-Id: I933cc3a70caecc17f5f50ae1d637528ee98cd9a9
> Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1546005
> Reviewed-by: Dirk Pranke <dpranke@chromium.org>
> Commit-Queue: Rakib Hasan <rmhasan@google.com>

TBR=dpranke@chromium.org,rmhasan@google.com

Change-Id: Iac3c7133b08291b6f0832399d182b011de0b744f
No-Presubmit: true
No-Tree-Checks: true
No-Try: true
Bug: chromium:698902