Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> I think I'm not sure what to look at; is patchset #1 the initial patch that was reverted, or does that include your fix?

Patchset #1 is not the original patch that was reverted. The changes is in line 564 or runner.py. It removes the loader.loadTestsFromName call and replaces it with setting the isolated_tests list of a TestSet instance to a list of TestInput objects.

Patch-set: 3
Reviewer: Gerrit User 1311766 <1311766@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>