Add /api/config
This API provides access to whitelisted configuration objects.
The only whitelisted object is revision_info for spa/details-table:
https://chromium-review.googlesource.com/c/catapult/+/1541649
Other objects may be whitelisted in the future, for example, to allow
teams to customize lists of test suites, sheriffs, or reports.

A followup will add a CacheRequest subclass to spa's service worker for
this API.

Bug: chromium:865515
Change-Id: Ie195b5b48df1dee9fb27f3a9770140a171d6239d