Pinpoint - Specify --gtest-benchmark-name flag for c++ tests
If this isn't specified, run_performance_test.py falls back to using
the benchmark (isolate) name from the command line. For some tests like
performance_browser_tests, the isolate/benchmark name differs and the
output is in a different location. By specifying the flag we both get
the output in the right place, and ensure we run right test.

Bug: chromium:870899
Change-Id: I48403adb4c7a86e9a329ea06a5590b29aa675eab