Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > This change is really cool, and I'm super interested in it.
> > 
> > For the devices that are on the chrome perf waterfall, we need to make sure we understand how much runtime increase this will cause. We don't currently run on the Pixel or Pixel 3, but we do run on the Pixel 2 and other other devices. If someone were to add support to this feature for the Pixel 2, then that could potentially cause a massive cycle time regression. I think we need some kind of guard against this where we make sure that anyone who tries to add this support will read a comment that will inform them of those performance implications. Preferably this guard code would be somewhere in Telemetry since we are the team that runs the waterfall rather than in Devil.
> 
> Note that we already do battery temperature checking and waiting elsewhere. It generally makes sense to accept the cycle time regression to let devices cool, both because we're less likely to damage the device and because we're less likely to be running under thermal throttling (which would adversely affect performance metrics).

Agree that this is likely the right direction. But we need to check for cycle time impact or otherwise we could easily demolish our entire speed infrastructure and start having timeouts everywhere. Some of this is tuning things correctly as well.

I haven't read the changelist in detail yet, but the commit message seems to say that this runs between every story. If if only runs before every benchmark, then that would be similar to what we already have with battery temperature.

Patch-set: 1
Reviewer: Gerrit User 1149061 <1149061@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>