devil: better error output for SetWebViewImplementation
No change to SetWebViewImplementation()'s overall functionality, but
this improves error output for failure cases.

Due to the complexity of loading WebView, SetWebViewImplementation() can
fail for a variety of (valid) reasons. Although the method correctly
detects success/failure, it only gives a single (vague) error message.
This modifies the method to give more precise (and actionable) error
messages.

As part of this implementation, this extends
GetWebViewUpdateServiceDump() to return more values in its dictionary.

Bug: chromium:949431
Test: ./devil/bin/run_py_tests
Test: out/Default/bin/system_webview_apk set-webview-provider (error
Test: cases are much clearer)
Change-Id: I83b8c78c6f4d5458bff0dcee4bc5d2ba25ae7ef6