[Telemetry] Replace some telemetry browser test logic with Typ's functions
We want to contain validation and removal of test name prefixes from test
names within Typ. In crrev.com/c/1544783, we added new functions that can
be used to add  test names to the test set, match test names to isolate
and skip globs, and filter tests using the test-filter argument. Each
function validates that the test name starts with the test name prefix and
then removes the prefix from the test name.

Bug: chromium:945481
Change-Id: Iafc4cd525a41617c19390c3403ea0b0671f17599