Update patch set 1
Patch Set 1:

When will this validation be done?

If it's done while running the tests - sorry, that's too late. It takes a really long time to run the WebGL 2.0 conformance tests, which is why the current validation that all of the test expectations actually refer to on-disk files is done as a presubmit check.

I have a feeling that this typ change isn't needed in order to preserve the current lint checking of the WebGL conformance expectation files.

Patch-set: 1
Reviewer: Gerrit User 1105344 <1105344@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>