[vinn] Raise original exception rather than exceptions hit during cleanup
We're getting the exception:

  File "c:\b\s\w\ir\third_party\catapult\third_party\vinn\vinn\_vinn.py", line 209, in RunFile
    _RemoveTreeWithRetry(temp_dir)
  File "c:\b\s\w\ir\third_party\catapult\third_party\vinn\vinn\_vinn.py", line 142, in _RemoveTreeWithRetry
    shutil.rmtree(tree)
  File "c:\b\s\w\ir\.swarming_module\bin\Lib\shutil.py", line 252, in rmtree
    onerror(os.remove, fullname, sys.exc_info())
  File "c:\b\s\w\ir\.swarming_module\bin\Lib\shutil.py", line 250, in rmtree
    os.remove(fullname)
WindowsError: [Error 32] The process cannot access the file because it is being used by another process: 'c:\\b\\s\\w\\it_u1w_m\\tmpyxj6rs\\_tmp_boostrap.js'


I suspect that this is caused by another exception downstream. However,
we can't see that original exception because we run this cleanup in
the "finally" clause. Instead, re-raise the original exception even
if you hit an error in the cleanup.

Bug: chromium:953365
Change-Id: I2967680250fd5eececaeb48bd268428faf534fa6