devil: better error output for invalid signature
This is a follow-up to http://crrev.com/c/1566774, to catch the case
where the WebView package has an "invalid" signature. This can happen if
no signature is specified in the WebView config (this is a
userdebug-only package) or if the package doesn't match the release-key
signature listed in the package (common, for local development).

I originally omitted this case because it only matters on user devices
(it's not enforced on userdebug/eng). But, there's no good reason to not
provide a clearer error message anyway (and a dedicated error message
makes this case easier to call out in developer guides anyway).

This shouldn't affect the method's behavior (it previously failed in
this case as well), but should just provide a clearer message.

Bug: chromium:955135
Test: ./devil/bin/run_py_tests
Test: out/Default/bin/monochrome_apk set-webview-provider (on user
Test: device)
Change-Id: I7cf089f0946638b83fac32db7c8ab7c9466a5c83