Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> (3 comments)

Done. The _GetThermalDeviceInformation is no longer called from the constructor. Instead it is used to intialise the _device_info variable. IsSupported() now checks to see if this variable is initialised and if not, it intialises it. 

Also, option 1 was implemented.

Patch-set: 2
Reviewer: Gerrit User 1338105 <1338105@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>