Update patch set 2
Patch Set 2:

The Trybots are failing for this CL because we're introducing a step to generate the proto in a subdirectory for the tests to run correctly. To make the trybots happy, we're going to either:

- Check-in the generated protobuf message (not recommended)
- Introduce a call to `make` in the `sheriff_config/` directory before running the tests

How do I accomplish the second option? Is there a recipe in the repository that I can tweak to introduce the call to `make` in the appropriate directory? Or does that change have to happen somewhere else?

Patch-set: 2
Reviewer: Gerrit User 1327821 <1327821@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0