Update patch set 5
Patch Set 5:

> Patch Set 4:
> 
> (2 comments)
> 
> Latest updated version:
> 
> https://dev-dberris-ef390ce0-dot-chromeperf.appspot.com/group_report?sid=7e6c9fca30be5712e4f56ed31c7dba0200152f99cd8f89575f9f0d0f3a6a9388

Thanks.
That link shows "141/192" in the expand button for the first group.
Before this change, the expand button showed the number of alerts in the group.
Now, I assume there are 192 alerts in the group since that's the bigger number, but the UI does not indicate what's different about 141 of them. Are there 141 alerts selected?
A tooltip like '141 untriaged out of 192 alerts in this group' might help a bit.
Alternatively, V2spa solves this problem by displaying two expand buttons in separate columns labeled "Count" and "Triaged", which also allows users to expand either just the triaged or just the untriaged alerts. V2spa also displays the number of selected alerts in the group next to the selection checkbox.

Most of the expand buttons show the same number of untriaged as total alerts. This is visually noisy, and might be cleaner if it only showed one of the numbers.

Patch-set: 5
Reviewer: Gerrit User 1115874 <1115874@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>