Update patch set 6
Patch Set 6: -Code-Review

PS1 is same as before, in PS6 I just revert back to my earlier approach to warn if tracing is not enabled.

I went a bit back and forth between options. It sounds like some tests can't enable tracing because they run as sub-processes of some other process that is already tracing. And Telemetry handles that by marking the agent as "not supported". So I thought of keeping the assert to ensure tracing is enabled if tracing is supported, and just log the fact that metadata is discarded otherwise. But then I thought this might cause other complications (e.g. not recording metadata when we should because tracing is not "controllable" although it actually *is* enabled).

To keep things simple I went back to my previous idea of just warning if the metadata is being discarded. What do you think?

TBR'ing benjhayden as there are no changes in tracing/* stuff.

Patch-set: 6
Reviewer: Gerrit User 1123010 <1123010@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0