Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > Patch Set 4:
> > 
> > (2 comments)
> > 
> > Latest updated version:
> > 
> > https://dev-dberris-ef390ce0-dot-chromeperf.appspot.com/group_report?sid=7e6c9fca30be5712e4f56ed31c7dba0200152f99cd8f89575f9f0d0f3a6a9388
> 
> Thanks.
> That link shows "141/192" in the expand button for the first group.
> Before this change, the expand button showed the number of alerts in the group.
> Now, I assume there are 192 alerts in the group since that's the bigger number, but the UI does not indicate what's different about 141 of them. Are there 141 alerts selected?
> A tooltip like '141 untriaged out of 192 alerts in this group' might help a bit.

Definitely -- I thought about doing this, but wasn't sure how to make that change easily. Based on my experience with tooltips in Polymer, getting dynamic content into those aren't as easy. I can add a "?" or help cue at the table header which will have the explanation, which could work as a substitute to a tooltip.

> Alternatively, V2spa solves this problem by displaying two expand buttons in separate columns labeled "Count" and "Triaged", which also allows users to expand either just the triaged or just the untriaged alerts. V2spa also displays the number of selected alerts in the group next to the selection checkbox.
> 
> Most of the expand buttons show the same number of untriaged as total alerts. This is visually noisy, and might be cleaner if it only showed one of the numbers.

I think the signal that all of the alerts in a group have been triaged is clearer if we point out that N/M means untriaged/total (by seeing both numbers in there). It reduces the confusion of how many of the alerts need attention.

This can get more complicated as we add more and more spreadsheet like features, which we can either punt on and instead export the data to Sheets to let folks slice and dice this the way they want. For example, I can imagine doing pivot tables instead, but aggregate based on different dimensions. That sounds like a waste of time compared to exporting the data as CSV and letting folks work on it how they want.

I'll send a follow-up to add the tooltip.

Patch-set: 5