Update patch set 4
Patch Set 4:

This is now just dependent on `protoc` being available in the builders. The CL to enable that is in https://chromium-review.googlesource.com/c/chromium/tools/build/+/1588024 -- once that lands we should be able to try this again, and see whether we can start generating the protocol buffers from the CI independent of whether we can use 'make' in Windows or not.

Patch-set: 4
Reviewer: Gerrit User 1327821 <1327821@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0