Update patch set 8
Patch Set 8:

I think we can fix this later, once we have more experience. It may be that we can change our minds later and because this is a protocol buffer, it's easy to change and let users adapt to the changes.

Landing now.

Patch-set: 8
Reviewer: Gerrit User 1327821 <1327821@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>